Split up test_plotting according to unit/integration tests
